backend/requirements.txt
structlog
starlette
pydantic-settings
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
alembic
pytest
httpx
